upload0.5

+++ src/app/components/vacation/vacation.component.ts
@@ -12,8 +12,19 @@ import { Router } from '@angular/router';
   templateUrl: './vacation.component.html',
   styleUrls: ['./vacation.component.css']
 })
-export class VacationComponent implements OnInit{
-  constructor(private _VacationService: VacationService, private _FormBuilder: FormBuilder , private _router:Router) { }
+export class VacationComponent implements OnInit {
+  constructor(private _VacationService: VacationService, private _FormBuilder: FormBuilder, private _router: Router) { }
+
+
+
+  empName: any = ''
+  empId: any = ''
+  availableDays: number = 0
+  takenDays: number = 0
+  requestSent: boolean = false
+  errorMaximumDays: boolean = false
+  todayDate: any = new Date().toISOString().split('T')[0]
+
 
   vacationForm: FormGroup = this._FormBuilder.group({
     startDate: [null, Validators.required], //  بداية الاجازة
@@ -22,40 +33,60 @@ export class VacationComponent implements OnInit{
     vacationType: [null, Validators.required],// نوع الاجازة 
   },)
 
-  empName: any = ''
-  empId: any = ''
-  availableDays:number = 0
-  takenDays:number = 0
-  todayDate: any = new Date().toISOString().split('T')[0]
 
 
 
+  sendRequest() {
 
+    console.log(this.vacationForm.value);
+    console.log(this.vacationForm.valid, 'valid');
+    if (this.vacationForm.get('numberOfDays')?.value <= this.availableDays && this.vacationForm.valid ) {
+        this._VacationService.requestvacation(this.vacationForm.value).subscribe({
+          next: (Response) => {
+            if(Response == true){
+              console.log(Response );
+              this.requestSent = true
+            }
 
+          },
 
+          error: (err) => {
+            console.log(err);
+            if (err.error.message == 'Unauthorized') {
+              localStorage.clear()
+              this._router.navigate(['login'])
+            }
+            this.requestSent = false
+          }
+        })
 
-
-
-
-ngOnInit(): void {
-  this._VacationService.basicvacationData().subscribe({
-    next: (Response) => {
-      this.empName = Response.name
-      this.empId = Response.empId
-      this.availableDays = Response.availableDays
-      this.takenDays = Response.takenDays
-      console.log(Response);
-    },
-
-    error: (err) => {
-      console.log(err);
-      if (err.error.message == 'Unauthorized') {
-        localStorage.clear()
-        this._router.navigate(['login'])
-      }
     }
-  })
-}
+    else{
+      this.vacationForm.markAllAsTouched()
+    }
+  }
+
+
+
+  ngOnInit(): void {
+    this._VacationService.basicvacationData().subscribe({
+      next: (Response) => {
+        this.empName = Response.name
+        this.empId = Response.empId
+        this.availableDays = Response.availableDays
+        this.takenDays = Response.takenDays
+        console.log(Response);
+      },
+
+      error: (err) => {
+        console.log(err);
+        if (err.error.message == 'Unauthorized') {
+          localStorage.clear()
+          this._router.navigate(['login'])
+        }
+      }
+    })
+  }
 
 
 }
